Support fractional factors and base amounts
This will now allow drawers to have fractional sizes instead of only whole sizes, removing the floor of the resultant size provider.
So for instance, now a drawer can store exactly 2 items by having a base size of (1/32) which is 0.04 (it's actually 0.03, but we use 0.04 because we don't round and we instead floor, so 0.03 would actually be only one item)

+++ src/main/java/com/buuz135/functionalstorage/block/tile/DrawerTile.java
@@ -38,7 +38,7 @@ public class DrawerTile extends ItemControllableDrawerTile<DrawerTile> {
             }
 
             @Override
-            public int getMultiplier() {
+            public float getMultiplier() {
                 return getStorageMultiplier();
             }
